mnist/other/requirements.txt
absl-py
astor==0.8.1
blinker==1.4
brotlipy==0.7.0
cachetools
certifi==2020.11.8
cffi
chardet
click==7.1.2
cryptography
cycler==0.10.0
docopt==0.6.2
Flask==1.1.2
gast==0.2.2
google-auth
google-auth-oauthlib
google-pasta==0.2.0
grpcio
h5py==2.10.0
idna
importlib-metadata
itsdangerous==1.1.0
Jinja2==2.11.2
Keras==2.3.1
Keras-Applications
Keras-Preprocessing==1.1.0
kiwisolver
MarkupSafe
matplotlib
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.1.0
Pillow
pipreqs==0.4.10
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser
pydot==1.4.1
PyJWT==1.7.1
pyOpenSSL
pyparsing==2.4.7
pyreadline==2.1
PySocks
python-dateutil==2.8.1
PyYAML==5.3.1
requests
requests-oauthlib==1.3.0
rsa
scipy
six
tensorboard
tensorboard-plugin-wit==1.6.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
tornado==6.0.4
urllib3
Werkzeug==0.16.1
win-inet-pton
wincertstore==0.2
wrapt==1.12.1
yarg==0.1.9
zipp
